função play movida para Game.js
